+++ src/frontend/components/layout/Banner.tsx
