+++ src/frontend/styles/pages/index.scss
@@ -1,2 +1,2 @@
-@use "matchLobby";
-
+@use "../match/matchLobby";
+@use "../match/playerBox";
