+++ src/pages/donations.tsx
@@ -1,0 +1,12 @@
+import Head from "next/head";
+
+export default function IndexPage() {
+  return (
+    <>
+      <Head>
+        <title>Support Wars World!</title>
+      </Head>
+      <h1>Work in progress...</h1>
+    </>
+  );
+}
