+++ src/pages/home/LoggedHome.tsx
@@ -1,0 +1,8 @@
+
+export default function LoggedHome() {
+  return (
+    <>
+      <h1>Work in progress</h1>
+    </>
+  );
+}
