+++ src/pages/index.tsx
@@ -1,15 +1,18 @@
 import Head from "next/head";
-import { Home } from "../frontend/components/Home";
+import BasicHome from "./home/BasicHome";
 
 export default function IndexPage() {
   return (
     <>
       <Head>
         <title>Home Page | Wars World</title>
-        <link rel="icon" href="/favicon.ico" />
       </Head>
-
-      <Home />
+      {
+        //TODO: Make it so there is a different home page
+        // when the user is logged in
+        //We'd have a basic home page and a specialized/logged in home page
+      }
+      <BasicHome />
     </>
   );
 }
