+++ src/pages/learn.tsx
@@ -1,0 +1,12 @@
+import Head from "next/head";
+
+export default function IndexPage() {
+  return (
+    <>
+      <Head>
+        <title>Wars World Guides</title>
+      </Head>
+      <h1>learning in progress...</h1>
+    </>
+  );
+}
