+++ src/pages/match/[matchId].tsx
@@ -9,9 +9,7 @@ export default function Match() {
         <title>Game Page | Wars World</title>
       </Head>
 
-      <Layout>
-        <PixiMatch />
-      </Layout>
+      <PixiMatch />
     </>
   );
 }
