+++ src/pages/terms.tsx
@@ -1,0 +1,12 @@
+import Head from "next/head";
+
+export default function IndexPage() {
+  return (
+    <>
+      <Head>
+        <title>WW Terms of Services</title>
+      </Head>
+      <h1>Work in progress...</h1>
+    </>
+  );
+}
