--- src/pages/terms/index.tsx
@@ -1,13 +1,0 @@
-import { Layout } from "frontend/components/layout";
-import Link from "next/link";
-
-export default function TermsPage() {
-  return (
-    <Layout>
-      <div className="@text-center @w-[500px]">
-        <div>To be constructed...</div>
-        <Link href="/">Home</Link>
-      </div>
-    </Layout>
-  );
-}
